+++ account_sdk/src/deploy_contract/mod.rs
@@ -3,3 +3,4 @@ mod deployment;
 mod pending;
 pub use declaration::CustomAccountDeclaration;
 pub use deployment::CustomAccountDeployment;
+pub use deployment::DeployResult;
