+++ account_sdk/src/tests/mod.rs
@@ -1,10 +1,3 @@
+mod deployment_test;
+mod erc20_methods_test;
 mod katana_test;
-mod new_deploy_test;
-mod one_file_all;
-
-#[macro_export]
-macro_rules! field_elem {
-    ($a:tt) => {
-        starknet::core::types::FieldElement::from_hex_be($a).unwrap()
-    };
-}
